requirements.txt
flask~=2.0.1
python-vlc~=3.0.21203
requests~=2.26.0
python-dotenv~=0.19.0 
werkzeug~=2.2.2
apscheduler~=3.9.1
gunicorn~=23.0.0
luma.oled
RPi.GPIO; sys_platform == 'linux' and platform_machine == 'aarch64'
spidev; sys_platform == 'linux' and platform_machine == 'aarch64'
